Some leaderboard functionality added.

+++ website/views.py
@@ -3,6 +3,7 @@ from .models import EquipDB, User
 from . import db
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import login_user, login_required, logout_user, current_user
+from datetime import date
 
 views = Blueprint('views', __name__)
 
@@ -11,7 +12,7 @@ views = Blueprint('views', __name__)
 def home():
     return render_template("home.html", user=current_user)
 
-#Allow user's to create equipment.
+#Page for users to create equipment.
 @views.route('/crafting', methods=["POST", "GET"])
 @login_required
 def crafting():
@@ -21,7 +22,7 @@ def crafting():
         equip = request.get_json()
 
         #Store the equipment information into the database.
-        equipment = EquipDB(equip)
+        equipment = EquipDB(equip, current_user.name, date.today())
         db.session.add(equipment)
         db.session.commit()
         print('Equipment information added to database.')
@@ -29,19 +30,19 @@ def crafting():
     else:
         return render_template("crafting.html", user=current_user)
 
-#Allow user's to view previously made equipment.
+#Page for user to view previously made equipment.
 @views.route('/crafts', methods=["POST", "GET"])
 @login_required
 def crafts():
 
-    saved_equips = EquipDB.query.all()
+    saved_equips = EquipDB.query.filter_by(userName=current_user.name)
     ordered_equips = []
 
     if request.method == "POST":
         #Removes equipment with given rowID from the database.
         if 'remove' in request.get_json()['methods']:
             rowID = request.get_json()['rowToDelete']
-            EquipDB.query.filter(EquipDB.id == rowID).delete()
+            EquipDB.query.filter_by(userName=current_user.name).filter(EquipDB.id == rowID).delete()
             db.session.commit()
             equips = EquipDB.query.all()
             
@@ -50,16 +51,27 @@ def crafts():
             colType = request.get_json()['sortType']
             sortType = request.get_json()['sortBy']
             if sortType == 'des':
-                equips = EquipDB.query.order_by(EquipDB.getAttribute(colType).desc()).all()
+                equips = EquipDB.query.filter_by(userName=current_user.name).order_by(EquipDB.getAttribute(colType).desc()).all()
             elif sortType == 'asc':
-                equips = EquipDB.query.order_by(EquipDB.getAttribute(colType).asc()).all()
+                equips = EquipDB.query.filter_by(userName=current_user.name).order_by(EquipDB.getAttribute(colType).asc()).all()
 
-        #Sends the database as a JSON object containing list of equipment to crafts.js to update HTML table.
+        #Returns the queried table as a JSON object containing list of equipment to crafts.js where it updates HTML table.
         for equip in equips:
             ordered_equips.append(equip.toJson())
         return jsonify(ordered_equips)
             
     return render_template("crafts.html", saved_equips=saved_equips, user=current_user)
+
+@views.route('/leaderboard', methods=["POST", "GET"])
+def leaderboard():
+    #TO DO:
+    #Datetime added to equipments database to track time
+    #Two leaderboards? One for daily, one for overall
+    #Click on column to sort by column (can use similar code from crafts.js)
+
+    saved_equips = EquipDB.query.order_by(EquipDB.getAttribute('gs').desc()).limit(10)
+
+    return render_template("leaderboard.html", saved_equips=saved_equips, user=current_user)
 
 @views.route('/register', methods=["POST", "GET"])
 def register():
@@ -68,7 +80,6 @@ def register():
         name = request.form['name']
         pw = request.form['pw']
         pw2 = request.form['pw2']
-        print(name, pw, pw2)
 
         #Form error handling
         if len(name) == 0 or len(pw) == 0 or len(pw2) == 0:
@@ -100,7 +111,6 @@ def login():
     if request.method == "POST":
         name = request.form['name']
         pw = request.form['pw']
-        print(name, pw)
 
         if len(name) == 0 or len(pw) == 0:
             flash('Please fill in all fields.', category='error')
